refactor: change package name for provider
