+++ main.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/providerserver"
 
-	"github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/cloudavenue"
+	"github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/provider"
 )
 
 // Example version string that can be overwritten by a release process.
@@ -27,7 +27,7 @@ func main() {
 		Debug:   debug,
 	}
 
-	err := providerserver.Serve(context.Background(), cloudavenue.New(version), opts)
+	err := providerserver.Serve(context.Background(), provider.New(version), opts)
 
 	if err != nil {
 		log.Fatal(err.Error())
